Created Framermotion config files
