+++ components/Masthead.Component.tsx
@@ -1,31 +1,8 @@
+import { stagger, fadeInUp } from "../src/utils/framer-motion/MastheadConfig";
 import { useRef, useContext, useState, useCallback } from "react";
 import { ScrollContext } from "../src/utils/scroll-observer";
-import { motion } from "framer-motion";
 import style from "../styles/Masthead.module.css";
-const easing = [0.6, -0.05, 0.01, 0.99];
-
-const fadeInUp = {
-  inital: {
-    y: 60,
-    opacity: 0,
-  },
-  animate: {
-    y: 0,
-    opacity: 1,
-    transition: {
-      duration: 0.6,
-      ease: easing,
-    },
-  },
-};
-
-const stagger = {
-  animate: {
-    transition: {
-      staggerChildren: 0.1,
-    },
-  },
-};
+import { motion } from "framer-motion";
 
 export const Masthead = () => {
   const [imageLoaded, setImageLoaded] = useState(true);
