+++ components/Status.Component.tsx
@@ -3,43 +3,11 @@ import Animation from "./Animation.Component";
 import { ApiDataState } from "../src/types/APIDataState";
 import style from "../styles/pages.module.css";
 import { motion } from "framer-motion";
-
-const easing = [0.6, -0.05, 0.01, 0.99];
-const fadeInUp = {
-  inital: {
-    y: 100,
-    opacity: 0,
-  },
-  animate: {
-    y: 0,
-    opacity: 1,
-    transition: {
-      duration: 0.6,
-      ease: easing,
-    },
-  },
-};
-const fadeInDown = {
-  inital: {
-    y: -100,
-    opacity: 0,
-  },
-  animate: {
-    y: 0,
-    opacity: 1,
-    transition: {
-      duration: 0.6,
-      ease: easing,
-    },
-  },
-};
-const stagger = {
-  animate: {
-    transition: {
-      staggerChildren: 0.1,
-    },
-  },
-};
+import {
+  stagger,
+  fadeInUp,
+  fadeInDown,
+} from "../src/utils/framer-motion/StatusConfig";
 
 interface Props {
   children: React.ReactNode;
